backend/requirements.txt
fastapi
uvicorn
torch
transformers
Pillow
python-multipart
opencv-python
pdf2image
